feat: added commands to interact with the configuration
* Added two commands: `list` and `set` to respectively
  list all the configuration keys and their values, and
  to update a value of the key.

Closes #42

+++ src/infrastructure/aoc_api/aoc_api_impl.rs
@@ -63,7 +63,7 @@ impl AocApi {
             .context("No answer in the parsed body")?;
         let answer_text = html2text::from_read(
             answer.text().collect::<Vec<_>>().join("").as_bytes(),
-            self.configuration.cli.output_width,
+            self.configuration.cli.output_width as usize,
         );
         Ok(answer_text)
     }
